SMS Notifications added
Currently linked to uvuslwcgamecenter@gmail.com
One click button to notify user of the station they are waiting for is ready.
Message is sent through all carriers to ensure the notification arrives.

Additionally phone number validation has been added to the entry field. Only numbers are accepted in this particular field when adding a user to the waitlist and there must be 10 digits.